requirements.txt
Gerekli Kütüphaneler:
customtkinter
openai
phyton-dotenv